requirements.txt
streamlit
opencv-python
dlib
imutils
numpy
Pillow
pygame
